:shirt: #26 split test file
